+++ Tests/TuistKitTests/Generator/TestData/ProjectDescription+TestData.swift
@@ -98,10 +98,10 @@ extension Scheme {
 }
 
 extension BuildAction {
-    static func test(targets: [String] = []) -> BuildAction {
-        BuildAction(targets: targets,
-                    preActions: [ExecutionAction.test()],
-                    postActions: [ExecutionAction.test()])
+    static func test(targets: [TargetReference] = []) -> BuildAction {
+        return BuildAction(targets: targets,
+                           preActions: [ExecutionAction.test()],
+                           postActions: [ExecutionAction.test()])
     }
 }
 
@@ -121,7 +121,7 @@ extension TestAction {
 
 extension RunAction {
     static func test(config: PresetBuildConfiguration = .debug,
-                     executable: String? = nil,
+                     executable: TargetReference? = nil,
                      arguments: Arguments? = nil) -> RunAction {
         RunAction(config: config,
                   executable: executable,
@@ -132,10 +132,10 @@ extension RunAction {
 extension ExecutionAction {
     static func test(title: String = "Test Script",
                      scriptText: String = "echo Test",
-                     target: String? = "Target") -> ExecutionAction {
-        ExecutionAction(title: title,
-                        scriptText: scriptText,
-                        target: target)
+                     target: TargetReference? = TargetReference(projectPath: nil, target: "Target")) -> ExecutionAction {
+        return ExecutionAction(title: title,
+                               scriptText: scriptText,
+                               target: target)
     }
 }